Deleted client-old, renamed variable to SourceVariable, fixed file naming
