--- client-old/.gitignore
@@ -1,8 +1,0 @@
-#Node
-node_modules
-
-#SASS Compilations
-*.css
-
-#Webstorm IDE
-.idea
