--- client-old/index.html
@@ -1,59 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <link rel="icon" href="images/favicon.png" type="image/x-icon" />
-    <!--Javascript-->
-    <script src="node_modules/angular/angular.js"></script>
-    <script src="node_modules/angular-animate/angular-animate.js"></script>
-    <script src="node_modules/angular-aria/angular-aria.js"></script>
-    <script src="node_modules/angular-material/angular-material.js"></script>
-    <script src="node_modules/angular-messages/angular-messages.js"></script>
-    <script src="https://use.fontawesome.com/bb80f3a922.js"></script>
-    <!--Spice Modules (These will be loaded dynamically for non -->
-    <script src="modules/spice.js"></script>
-    <script src="modules/Stream.js"></script>
-    <script src="modules/SpiceError.js"></script>
-    <script src="modules/filesystem-service.js"></script>
-    <script src="modules/debugger-service.js"></script>
-    <script src="modules/launcher/launcher.js"></script>
-    <script src="modules/functionPicker/functionPicker.js"></script>
-    <script src="modules/debugViewer/debugViewer.js"></script>
-    <!--CSS-->
-    <link rel="stylesheet" href="node_modules/angular-material/angular-material.css">
-    <link rel="stylesheet" href="spice.css">
-    <title>(Old) Spice Debug</title>
-</head>
-<body ng-app="Spice" ng-controller="MainCtrl as ctrl" ng-cloak md-theme="docs-dark">
-    <md-content layout-fill>
-        <md-toolbar>
-            <div class="md-toolbar-tools">
-                <h1 style="position:absolute"><b>Spice Debugger</b></h1>
-                <span flex></span>
-                <md-button ng-disabled="ctrl.view == 'launcher'" ng-click="ctrl.view = 'launcher'" aria-label="Launcher">
-                    Launcher
-                </md-button>
-                <i class="fa fa-arrow-right" aria-hidden="true"></i>
-                <md-button ng-disabled="ctrl.view == 'configuration' || !ctrl.getConfigUnlocked()" ng-click="ctrl.view = 'configuration'" aria-label="Configuration">
-                    Configuration
-                </md-button>
-                <i class="fa fa-arrow-right" aria-hidden="true"></i>
-                <md-button ng-disabled="ctrl.view == 'debugging' || !ctrl.getDebuggingUnlocked()" ng-click="ctrl.view = 'debugging'" aria-label="Debugger">
-                    Debugging
-                </md-button>
-                <span flex></span>
-                <a href="https://github.com/team-worm/spice" target="_blank"><i class="fa fa-github" aria-hidden="true"></i></a>
-            </div>
-        </md-toolbar>
-        <md-progress-linear md-mode="{{Loader.mode}}" class="md-warn" value="0"></md-progress-linear>
-        <spice-launcher mode="mode" ng-if="ctrl.view == 'launcher'" loader="Loader" ng-cloak></spice-launcher>
-        <spice-function-picker ng-if="ctrl.view == 'configuration'" loader="Loader" ng-cloak></spice-function-picker>
-        <spice-debug-viewer ng-if="ctrl.view == 'debugging'" ng-cloak></spice-debug-viewer>
-        <div class="mockup">
-            <span>
-                Prototype
-            </span>
-        </div>
-    </md-content>
-</body>
-</html>
