--- client-old/modules/debugViewer/debugViewerTemplate.html
@@ -1,26 +1,0 @@
-<h3>Output of binarySearch(int, int*, int)</h3>
-<md-content ng-controller="DebugViewerCtrl" class="debugger">
-    <reactive-height-grid layout="row">
-        <div flex class="source-code-viewer">
-            <div layout="row">
-                <div flex="none" class="line-number-column debug-column">
-                    <reactive-height-cell ng-repeat="line in lines" row=$index col="0">
-                        <span flex>{{$index + 1}}</span>
-                    </reactive-height-cell>
-                </div>
-                <div flex class="code-column debug-column">
-                    <reactive-height-cell ng-repeat="line in lines" row="$index" col="1">
-                        <pre flex>{{line.code}}</pre>
-                    </reactive-height-cell>
-                </div>
-            </div>
-        </div>
-        <div flex class="state-column debug-column" layout="row">
-            <div flex class="trace-column" ng-repeat="(colIndex, i) in [].constructor(traceColCount) track by $index">
-                <reactive-height-cell ng-repeat="line in lines" row="$index" col="$parent.$index+2">
-                    <trace flex trace="line.traces[colIndex]"></trace>
-                </reactive-height-cell>
-            </div>
-        </div>
-    </reactive-height-grid>
-</md-content>
