--- client-old/modules/debugViewer/traceTemplate.html
@@ -1,14 +1,0 @@
-<div class="trace">
-    <div ng-if="trace.tType===0" class="trace-state">
-        <ul>
-            <li ng-repeat="state in trace.data">{{getVariable(state.variable).name}}={{state.value}}</li>
-        </ul>
-    </div>
-    <div ng-if="trace.tType===1" class="trace-output">
-        output: {{trace.data.output}}
-    </div>
-    <div ng-if="trace.tType===2" class="trace-terminated">
-        <!-- terminated: {{trace.data.cause}} -->
-        returned 4
-    </div>
-</div>
