--- client-old/modules/functionPicker/functionPickerTemplate.html
@@ -1,72 +1,0 @@
-<md-content ng-controller="FunctionPickerCtrl as ctrl" class="configuration">
-    <md-button class="md-raised md-primary"
-               ng-disabled="!ctrl.getCanRunFunction()"
-               ng-click="ctrl.runFunction()" aria-label="Help with Launch a Binary">
-        Run Function <i class="fa fa-play" aria-hidden="true"></i>
-    </md-button>
-    <md-button class="md-raised md-primary" ng-disabled="!ctrl.getCanKillFunction()" ng-click="ctrl.killFunction()"
-               aria-label="Help with Launch a Binary">
-        Kill Running Function <i class="fa fa-window-close" aria-hidden="true"></i>
-    </md-button>
-    <span class="error-text">{{ctrl.error}}</span>
-    <hr>
-    <md-content class="md-padding">
-        <md-toolbar class="md-accent">
-            <div class="md-toolbar-tools">
-                <b>binary-search.exe</b>
-                <i class="spacer fa fa-caret-right" aria-hidden="true"> </i>
-                <b>binary-search.c</b>
-                <i class="spacer fa fa-caret-right"  ng-show="!!ctrl.selectedFunction" aria-hidden="true"> </i>
-                <b>{{ctrl.selectedFunction}}</b>
-            </div>
-        </md-toolbar>
-        <md-content flex="100" layout="row" layout-xs="column" class="md-padding md-whiteframe-3dp">
-            <md-content flex="30" flex-xs="100" class="md-padding configuration">
-                <md-autocomplete
-                        flex
-                        md-items="item in ctrl.FunctionList"
-                        md-min-length="0"
-                        md-selected-item="ctrl.selectedFunction"
-                        placeholder="Search a File or Function ">
-                    <md-item-template>
-                        <span>{{item}}</span>
-                    </md-item-template>
-                    <md-not-found>
-                        Function or File Not Found!
-                    </md-not-found>
-                </md-autocomplete>
-                <md-virtual-repeat-container id="function-list">
-                    <div md-virtual-repeat="item in ctrl.FunctionList"
-                         class="repeated-item"
-                         ng-click="ctrl.setFunctionToRun(item)">
-                        <span>{{item}}</span>
-                    </div>
-                </md-virtual-repeat-container>
-                <md-content id="console" layout-padding>
-                    <pre> > I am a console.
- > I will print output of the running function.
- > This will be functional in the final product.
- > ...</pre>
-                </md-content>
-            </md-content>
-            <md-content flex="70" flex-xs="100" class="md-whiteframe-3dp debugger">
-                <reactive-height-grid layout="row">
-                    <div flex class="source-code-viewer">
-                        <div layout="row">
-                            <div flex="none" class="line-number-column debug-column">
-                                <reactive-height-cell ng-repeat="line in ctrl.lines" row=$index col=0>
-                                    <span flex>{{$index + 1}}</span>
-                                </reactive-height-cell>
-                            </div>
-                            <div flex class="code-column debug-column">
-                                <reactive-height-cell ng-repeat="line in ctrl.lines" row=$index col=1>
-                                    <pre flex>{{line.code}}</pre>
-                                </reactive-height-cell>
-                            </div>
-                        </div>
-                    </div>
-                </reactive-height-grid>
-            </md-content>
-        </md-content>
-    </md-content>
-</md-content>
