--- client-old/package.json
@@ -1,28 +1,0 @@
-{
-  "name": "client",
-  "version": "1.0.0",
-  "description": "Debug through time.",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "sass": "node node_modules/node-sass/bin/node-sass styles/spice.scss spice.css",
-    "sass-watch": "node node_modules/node-sass/bin/node-sass --watch styles/spice.scss spice.css",
-    "serve": "http-server -p 8001 -c-1"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "angular": "^1.5.8",
-    "angular-animate": "^1.5.8",
-    "angular-aria": "^1.5.8",
-    "angular-material": "^1.1.1",
-    "angular-messages": "^1.5.8"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/team-worm/spice.git"
-  },
-  "devDependencies": {
-    "node-sass": "^3.13.0"
-  }
-}
