--- client-old/styles/spice.scss
@@ -1,143 +1,0 @@
-$background-color-0: #2B2B2B;
-$background-color-1: #313335;
-$background-color-2: #3C3F41;
-$background-color-3: #ca934d;
-
-$foreground-color-0: #B5B5B5;
-
-$foreground-color-2: #dddddd;
-$foreground-color-3: #ffffff;
-
-$highlight-primary-color-0: #feb75b;
-$highlight-secondary-color-0: #2069c7;
-
-$background-highlight-primary-color-0: #715034;
-$background-highlight-secondary-color-0: #b26f34;
-
-$highlight-alt-color-0: rgb(233, 30, 99);
-
-$text-color-error: rgb(220, 49, 32);
-
-.mockup {
-    color: gray;
-    padding-top: 1em;
-    width: 100%;
-    text-align: center;
-    font-size: larger;
-}
-
-.error-text {
-    color: $text-color-error;
-    font-style:italic;
-}
-
-.launcher {
-    .bottom-right {
-        position: absolute;
-        bottom: 0;
-        right: 0;
-    }
-
-    .error-box {
-        text-align: center;
-    }
-
-    .filesystem {
-        :focus {
-            outline: 0;
-        }
-        .file {
-            &:hover {
-                cursor: pointer;
-                color: black;
-                background-color: $foreground-color-2;
-            }
-            &.lit {
-                color: black;
-                padding: 2px;
-                border: 2px solid $highlight-secondary-color-0;
-                background-color: $foreground-color-2;
-            }
-        }
-    }
-}
-
-.configuration {
-
-    .spacer {
-        margin-left:1em;
-        margin-right:1em;
-    }
-    #console {
-        background: black;
-        color: white;
-        margin-top: 1em;
-        font-family: Courier, 'NewCourier', monospace;
-        border: 1px solid gray;
-    }
-
-    #function-list {
-        height: 6em;
-        width: 100%;
-        border: 1px solid $highlight-alt-color-0;
-        margin-top: 1em;
-
-        .repeated-item {
-            border-bottom: 1px solid #aaa;
-            cursor: pointer;
-            padding-top: 10px;
-            padding-left: 5px;
-            height: 2em;
-        }
-    }
-}
-
-.debugger {
-    background: $background-color-0;
-    font-family: Courier, 'NewCourier', monospace;
-    font-size: 1em;
-
-    .reactive-height-cell {
-
-        padding-left: 8px;
-        padding-right: 8px;;
-
-        background: $background-color-0;
-
-        pre {
-            margin: 2px;
-        }
-    }
-
-    .debug-column reactive-height-cell:nth-of-type(odd) .reactive-height-cell {
-        background: $background-color-1;
-    }
-
-    .line-number-column {
-        background: $background-color-1;
-        border-left: 2px solid $background-color-2;
-        border-right: 2px solid $background-color-2;
-        text-align: right;
-    }
-
-    .code-column {
-        border-right: 4px solid $background-color-2;
-    }
-
-    .state-column {
-        .trace-column {
-            border-right: 2px solid $background-color-2;
-            .trace {
-                .trace-state {
-                    ul {
-                        margin: 0px;
-                        padding-left: 0px;
-                    }
-                    li {
-                        list-style-type: none;
-                    }
-                }
-            }
-        }
-    }
-}
