+++ client/app/components/common/file-browser.component.ts
