+++ client/app/models/SourceFunction.ts
@@ -1,4 +1,4 @@
-import {Variable} from "./Variable";
+import {SourceVariable} from "./SourceVariable";
 
 export class SourceFunction {
     public address: number; //Memory address of the function.
@@ -6,6 +6,6 @@ export class SourceFunction {
     public sourcePath: string; //Full path to the source code.
     public lineNumber: number; //Line number of the function definition.
     public lineCount: number; //Length of function in source code lines.
-    public parameters: Variable[]; //Array of parameter variables.
-    public localVariables: Variable[]; //Array of all local variables defined in function.
+    public parameters: SourceVariable[]; //Array of parameter variables.
+    public localVariables: SourceVariable[]; //Array of all local variables defined in function.
 }
