+++ client/app/models/SourceVariable.ts
@@ -5,7 +5,7 @@
 import {SourceType} from "./SourceType";
 
 
-export class Variable {
+export class SourceVariable {
     public id: string; //Unique identifier of that variable.
     public name: string; //Name of the variable as it appears in the function.
     public sType: SourceType; //Type of the variable as it is defined in the source code.
