+++ client/app/spice.module.ts
@@ -4,7 +4,7 @@ import {BrowserModule} from '@angular/platform-browser';
 import {MaterialModule, MdIconRegistry} from "@angular/material";
 import {FormsModule} from "@angular/forms";
 
-import {SpiceRootComponent} from "./components/spice.root.component";
+import {SpiceRootComponent} from "./components/spice-root.component";
 import {ToolbarComponent} from "./components/toolbar/toolbar.component";
 import {AboutComponent} from "./components/toolbar/about.component";
 import {HelpComponent} from "./components/toolbar/help.component";
@@ -12,7 +12,7 @@ import {LauncherComponent} from "./components/launcher/launcher.component";
 import {ConfigurationComponent} from "./components/configuration/configuration.component";
 import {DebuggerComponent} from "./components/debugger/debugger.component";
 
-import {FileSystemService} from "./services/file.system.service";
+import {FileSystemService} from "./services/file-system.service";
 import {ViewService} from "./services/view.service";
 
 @NgModule({
